requirements.txt
flask
requests
openai
python-dotenv
gspread
oauth2client
waitress